Relocate the code in preparation for future enhancement
